requirements.txt
Adafruit_DHT>=1.3.4
pygame>=1.9.3
requests>=2.19.1
RPi.GPIO>=0.6.3
